avatar page and edit page is configured
