+++ app/routes/_main._general.avatars.($id).edit.tsx
@@ -8,8 +8,11 @@ import * as Input from '~/components/ui/input/input';
 import * as Textarea from '~/components/ui/input/textarea';
 import PlayerButton from '~/components/PlayerButton';
 import { PATHS } from '~/constants';
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import SelectVoiceModal from '~/components/selectVoiceModal';
+import { cn } from '~/utils/cn';
+import { getPicture } from '~/utils/getPicture';
+import PublishAvatarModal from '~/components/publishAvatarModal';
 
 export function meta({}: Route.MetaArgs) {
   return [{ title: 'Avatar edit' }];
@@ -35,9 +38,45 @@ export default function AvatarEdit({ loaderData }: Route.ComponentProps) {
   const ttsVoices: TtsVoice[] = loaderData.ttsVoices;
   const fetcher = useFetcher();
   const [selectedVoice, setSelectedVoice] = useState<TtsVoice>(avatar.ttsVoice);
+  const [selectedImage, setSelectedImage] = useState<string | null>(avatar.picture ?? null);
+  const [preventFileOpen, setPreventFileOpen] = useState(false);
+  const fileInputRef = useRef<HTMLInputElement | null>(null);
+  const [availability, setAvailability] = useState<'private' | 'public'>(avatar.published ? 'public' : 'private');
 
   const handleVoiceChange = (voice: TtsVoice) => {
     setSelectedVoice(voice);
+  };
+
+  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (e.target.files && e.target.files[0]) {
+      const file = e.target.files[0];
+      const imageUrl = URL.createObjectURL(file);
+      setSelectedImage(imageUrl);
+    }
+  };
+
+  const handleLabelClick = (e: React.MouseEvent) => {
+    if (preventFileOpen) {
+      e.preventDefault();
+      setPreventFileOpen(false);
+      return;
+    }
+  };
+
+  const handleTrashClick = (e: React.MouseEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setSelectedImage(null);
+
+    setPreventFileOpen(true);
+
+    if (fileInputRef.current) {
+      fileInputRef.current.value = '';
+    }
+  };
+
+  const handlePublishConfirm = () => {
+    setAvailability('public');
   };
 
   return (
@@ -97,6 +136,48 @@ export default function AvatarEdit({ loaderData }: Route.ComponentProps) {
             </div>
           </div>
           <div className='sm:pl-4 sm:max-w-[352px] flex size-full flex-col gap-10'>
+            <div className='relative'>
+              <label
+                className='sm:h-60 h-[263px] w-full bg-none sm:bg-transparent bg-neutral-04 sm:bg-[linear-gradient(86.23deg,rgba(254,253,248,0.56)_0%,rgba(255,255,255,0.56)_100%)] sm:backdrop-blur-48 flex flex-col justify-end items-center gap-3.5 rounded-xl cursor-pointer relative'
+                onClick={handleLabelClick}
+              >
+                <input ref={fileInputRef} className='hidden' type='file' name='picture' accept='image/*' onChange={handleImageChange} />
+                {selectedImage !== null ? (
+                  <div className='size-full'>
+                    <img
+                      src={selectedImage.startsWith('blob:') ? selectedImage : getPicture(avatar, 'avatars', false)}
+                      srcSet={!selectedImage.startsWith('blob:') ? getPicture(avatar, 'avatars', true) : undefined}
+                      alt={avatar.name}
+                      className='size-full object-cover rounded-lg'
+                    />
+                  </div>
+                ) : (
+                  <div className='flex items-center justify-center size-full'>
+                    <Icons.fileUploadIcon />
+                  </div>
+                )}
+              </label>
+              <div className='absolute bottom-3 left-3'>
+                <PlayerButton variant='white' className='shadow-bottom-level-1' audioSrc={PATHS.ttsVoice(avatar.ttsVoiceId)} />
+              </div>
+              <div className='absolute z-10 bottom-3 right-3'>
+                <div className='flex items-center justify-between w-full'>
+                  <div
+                    className={cn(
+                      'py-2 px-5 flex items-center justify-center bg-base-white shadow-bottom-level-1 rounded-full',
+                      (selectedImage || avatar.picture) && 'divide-x divide-neutral-04 gap-4'
+                    )}
+                  >
+                    {selectedImage !== null && (
+                      <button type='button' className='pr-4 relative z-10' onClick={handleTrashClick}>
+                        <Icons.trash className='text-black' />
+                      </button>
+                    )}
+                    <Icons.fileUpload className='cursor-pointer' onClick={() => fileInputRef.current?.click()} />
+                  </div>
+                </div>
+              </div>
+            </div>
             <div className='flex flex-col gap-5'>
               <div className='flex items-center justify-between'>
                 <h1 className='text-base-black text-heading-h3 font-semibold'>Voice</h1>
@@ -111,6 +192,58 @@ export default function AvatarEdit({ loaderData }: Route.ComponentProps) {
                 </div>
               </div>
             </div>
+            <div className='sm:flex hidden flex-col gap-5'>
+              <h1 className='text-base-black text-heading-h3 font-semibold sm:block hidden'>Availability</h1>
+              <div className='p-1 bg-none sm:bg-transparent bg-neutral-04 sm:bg-[linear-gradient(86.23deg,rgba(254,253,248,0.56)_0%,rgba(255,255,255,0.56)_100%)] grid grid-cols-2 rounded-xl'>
+                <button
+                  type='button'
+                  className={cn(
+                    'flex items-center justify-center py-3 text-body-sm font-semibold rounded-xl transition-colors',
+                    availability === 'private' ? 'bg-white' : 'bg-transparent'
+                  )}
+                  onClick={() => setAvailability('private')}
+                >
+                  🔒 Private
+                </button>
+                <PublishAvatarModal onConfirm={handlePublishConfirm}>
+                  <button
+                    type='button'
+                    className={cn(
+                      'flex items-center justify-center py-3 text-body-sm font-semibold rounded-xl transition-colors',
+                      availability === 'public' ? 'bg-white' : 'bg-transparent'
+                    )}
+                  >
+                    🌐 Public
+                  </button>
+                </PublishAvatarModal>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div className='flex sm:hidden flex-col gap-5 sm:static absolute w-full sm:w-auto left-0 px-4.5 pt-4.5 sm:px-0 sm:pt-0 sm:pb-0 pb-4.5 sm:rounded-t-none rounded-t-xl bottom-0 sm:bg-none bg-[linear-gradient(86.23deg,rgba(254,253,248,0.48)_0%,rgba(255,255,255,0.48)_100%)] shadow-bottom-bar backdrop-blur-48'>
+          <h1 className='text-base-black text-heading-h3 font-semibold sm:block hidden'>Availability</h1>
+          <div className='p-1 bg-none sm:bg-transparent bg-neutral-04 sm:bg-[linear-gradient(86.23deg,rgba(254,253,248,0.56)_0%,rgba(255,255,255,0.56)_100%)] grid grid-cols-2 rounded-xl'>
+            <button
+              type='button'
+              className={cn(
+                'flex items-center justify-center py-3 text-body-sm font-semibold rounded-xl transition-colors',
+                availability === 'private' ? 'bg-white' : 'bg-transparent'
+              )}
+              onClick={() => setAvailability('private')}
+            >
+              🔒 Private
+            </button>
+            <PublishAvatarModal onConfirm={handlePublishConfirm}>
+              <button
+                type='button'
+                className={cn(
+                  'flex items-center justify-center py-3 text-body-sm font-semibold rounded-xl transition-colors',
+                  availability === 'public' ? 'bg-white' : 'bg-transparent'
+                )}
+              >
+                🌐 Public
+              </button>
+            </PublishAvatarModal>
           </div>
         </div>
       </div>
